+++ src/test/java/redis/clients/jedis/csc/JedisPooledClientSideCacheTest.java
@@ -20,7 +20,6 @@ import redis.clients.jedis.HostAndPorts;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisClientConfig;
 import redis.clients.jedis.JedisPooled;
-import redis.clients.jedis.csc.util.MapCSC;
 
 public class JedisPooledClientSideCacheTest {
